Abstract HTTP logic to a chatbot-service
