+++ .prettierrc
@@ -1,6 +1,6 @@
 {
   "trailingComma": "all",
-  "endOfLine": "lf",
+  "endOfLine": "auto",
   "tabWidth": 2,
   "semi": true,
   "singleQuote": true
