+++ chatbot-ui/hooks/useChatbot.ts
@@ -1,42 +1,5 @@
 import { useState } from 'react';
-
-const api_url =
-  process.env.NEXT_PUBLIC_CHAT_URL_API ?? 'http://localhost:8010/api/chat';
-
-const strangle_api = process.env.NEXT_TEMP_STRANGLE_API ?? true;
-const delay = (ms: number) => new Promise((res) => setTimeout(res, ms));
-
-const parseResponseFromAPI = async (response: Response): Promise<Message> => {
-  const data = response.body;
-
-  if (!data) {
-    throw new Error('Response contained no body.');
-  }
-
-  const reader = data.getReader();
-  const decoder = new TextDecoder();
-  let done = false;
-
-  let content = '';
-  let links = undefined;
-
-  while (!done) {
-    const { value, done: doneReading } = await reader.read();
-    done = doneReading;
-    const chunkValue = decoder.decode(value);
-    if (chunkValue.startsWith('{"sourceDocuments":')) {
-      links = JSON.parse(chunkValue).sourceDocuments;
-    } else {
-      content += chunkValue;
-    }
-  }
-
-  return {
-    content,
-    links,
-    type: 'apiMessage',
-  };
-};
+import sendUserMessage from '@/services/chatbot-service';
 
 function useChatbot(): UseChatbotState {
   const [fetching, setFetching] = useState<boolean>(false);
@@ -49,40 +12,20 @@ function useChatbot(): UseChatbotState {
     },
   ]);
 
-  const recordMessageInHistory = (message: Message) => {
-    setMessageHistory(messageHistory.concat([message]));
+  const recordMessageInHistory = (messages: Message[]) => {
+    setMessageHistory(messageHistory.concat(messages));
   };
 
   const sendMessage = async (userInput: string) => {
-    recordMessageInHistory({ content: userInput, type: 'userMessage' });
+    const userMessage: Message = { content: userInput, type: 'userMessage' };
+    recordMessageInHistory([userMessage]);
     setFetching(true);
 
-    if (strangle_api) {
-      await delay(5000);
-
-      const fakeMessage: Message = {
-        content: 'This is a fake mocked out message',
-        type: 'apiMessage',
-      };
-      recordMessageInHistory(fakeMessage);
-      setFetching(false);
-      return;
-    }
-
     try {
-      const response = await fetch(api_url, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          userInput,
-        }),
-      });
+      const responseMessage = await sendUserMessage(userInput);
 
-      const responseMessage = await parseResponseFromAPI(response);
-
-      recordMessageInHistory(responseMessage);
+      // TODO: Refactor state management so this works without having to re-add user message
+      recordMessageInHistory([userMessage, responseMessage]);
     } catch (error) {
       setError('An error occurred while fetching the data. Please try again.');
     }
